New Class called guess and improved alpha
Full working alphaversion
